Jb53610 unit test improvements

+++ tests/common.cpp
@@ -129,6 +129,70 @@ void waitForSignal(QObject *object, const char *signal)
     loop.exec();
 }
 
+ContactChangeListener::ContactChangeListener()
+{
+    m_loop = new QEventLoop;
+    m_timer = new QTimer;
+    SeasideCache::registerChangeListener(this);
+}
+
+ContactChangeListener::~ContactChangeListener()
+{
+    delete m_loop;
+    delete m_timer;
+    SeasideCache::unregisterChangeListener(this);
+}
+
+bool ContactChangeListener::waitForContactAdded(int contactId)
+{
+    return waitForChange(contactId, &m_waitForContactAddedId, &m_updatedContactIds);
+}
+
+bool ContactChangeListener::waitForContactDeleted(int contactId)
+{
+    return waitForChange(contactId, &m_waitForContactDeletedId, &m_deletedContactIds);
+}
+
+bool ContactChangeListener::waitForChange(quint32 contactId, QSet<quint32> *changeWaits, QSet<quint32> *changeRecords)
+{
+    if (changeRecords->contains(contactId)) {
+        qDebug() << "Change listener already contains target:" << contactId;
+        return true;
+    }
+
+    changeWaits->insert(contactId);
+
+    m_timer->start(10 * 1000);
+    QObject::connect(m_timer, &QTimer::timeout, m_loop, &QEventLoop::quit);
+    m_loop->exec();
+
+    const bool success = changeWaits->contains(contactId);
+    changeWaits->remove(contactId);
+    changeRecords->remove(contactId);
+
+    return success;
+}
+
+void ContactChangeListener::itemUpdated(SeasideCache::CacheItem *item)
+{
+    const quint32 contactId = SeasideCache::internalId(item->contact);
+    m_updatedContactIds.insert(contactId);
+
+    if (m_waitForContactAddedId.contains(contactId) && m_loop->isRunning()) {
+        m_loop->quit();
+    }
+}
+
+void ContactChangeListener::itemAboutToBeRemoved(SeasideCache::CacheItem *item)
+{
+    const quint32 contactId = SeasideCache::internalId(item->contact);
+    m_deletedContactIds.insert(contactId);
+
+    if (m_waitForContactDeletedId.contains(contactId) && m_loop->isRunning()) {
+        m_loop->quit();
+    }
+}
+
 }
 
 using namespace CommHistory;
@@ -145,9 +209,8 @@ QSet<int> addedEventIds;
 
 void initTestDatabase()
 {
-    deleteAll();
-
     CommHistoryDatabasePath::setRootDir(TEST_DATABASE_DIR);
+    deleteAll();
 }
 
 int addTestEvent(EventModel &model,
@@ -210,7 +273,7 @@ void addTestGroup(Group& grp, QString localUid, QString remoteUid)
     QSignalSpy ready(&groupModel, SIGNAL(groupsCommitted(QList<int>,bool)));
     QVERIFY(groupModel.addGroup(grp));
 
-    QVERIFY(waitSignal(ready));
+    QTRY_COMPARE(ready.count(), 1);
     QVERIFY(ready.first().at(1).toBool());
 }
 
@@ -223,7 +286,7 @@ QContactId localContactForAggregate(const QContactId &aggregateId)
     return QContactId();
 }
 
-int addTestContact(const QString &name, const QString &remoteUid, const QString &localUid)
+int addTestContact(const QString &name, const QString &remoteUid, const QString &localUid, ContactChangeListener *listener)
 {
     QString contactUri = QString("<testcontact:%1>").arg(contactNumber++);
 
@@ -236,14 +299,25 @@ int addTestContact(const QString &name, const QString &remoteUid, const QString 
         return -1;
     }
 
+    int retValue = -1;
     foreach (const QContactRelationship &relationship, manager()->relationships(QContactRelationship::Aggregates(), contact.id(), QContactRelationship::Second)) {
         const QContactId &aggId = relationship.first();
         addedContactIds.insert(aggId);
-        return internalContactId(aggId);
+        retValue = internalContactId(aggId);
+        break;
     }
 
-    qWarning() << "Could not find aggregator";
-    return internalContactId(contact.id());
+    if (retValue < 0) {
+        qWarning() << "Could not find aggregator";
+        retValue = internalContactId(contact.id());
+    }
+
+    if (listener && !listener->waitForContactAdded(retValue)) {
+        qWarning() << "Test contact" << retValue << "was not added!";
+        return -1;
+    }
+
+    return retValue;
 }
 
 QList<int> addTestContacts(const QList<QPair<QString, QPair<QString, QString> > > &details)
@@ -329,8 +403,6 @@ bool addTestContactAddress(int contactId, const QString &remoteUid, const QStrin
 
 void modifyTestContact(int id, const QString &name, bool favorite)
 {
-    qDebug() << Q_FUNC_INFO << id << name;
-
     QContact existingAggregate = manager()->contact(apiContactId(id, manager()->managerUri()));
     if (internalContactId(existingAggregate.id()) != (unsigned)id) {
         qWarning() << "Could not retrieve contact:" << id;
@@ -363,7 +435,7 @@ void modifyTestContact(int id, const QString &name, bool favorite)
     }
 }
 
-void deleteTestContact(int id)
+void deleteTestContact(int id, ContactChangeListener *listener)
 {
     const QContactId contactId = apiContactId(id, manager()->managerUri());
     const QContact contact = manager()->contact(contactId);
@@ -375,6 +447,10 @@ void deleteTestContact(int id)
         qWarning() << "error deleting contact:" << contactId.localId();
     }
     addedContactIds.remove(contactId);
+
+    if (listener && !listener->waitForContactDeleted(SeasideCache::internalId(contactId))) {
+        qWarning() << "Test contact" << contactId << "was not deleted!";
+    }
 }
 
 void cleanupTestGroups()
@@ -484,12 +560,14 @@ bool compareEvents(Event &e1, Event &e2)
     return true;
 }
 
-void deleteAll()
+void deleteAll(bool deleteDb)
 {
-    qDebug() << Q_FUNC_INFO << "- Deleting all";
-
     cleanupTestGroups();
     cleanupTestEvents();
+
+    if (!deleteDb) {
+        return;
+    }
 
     if (!QDir(TEST_DATABASE_DIR).removeRecursively()) {
         qWarning() << "Unable to remove test database directory:" << TEST_DATABASE_DIR;
@@ -512,22 +590,6 @@ QString randomMessage(int words)
         msgStream << msgWords[qrand() % numWords] << " ";
     }
     return msg;
-}
-
-bool waitSignal(QSignalSpy &spy, int msec)
-{
-    if (!spy.isEmpty()) {
-        return true;
-    }
-    QElapsedTimer timer;
-    timer.start();
-    while (timer.elapsed() < msec && spy.isEmpty()) {
-        QCoreApplication::sendPostedEvents();
-        QCoreApplication::sendPostedEvents(0, QEvent::DeferredDelete);
-        QCoreApplication::processEvents();
-    }
-
-    return !spy.isEmpty();
 }
 
 void summarizeResults(const QString &className, QList<int> &times, QFile *logFile, int testSecs)
